Davide's edits
The intro was only on sourceforge, I append it as a note. Depending on the target, we probably need some introduction.

I added a section "Advice based on a constrain/driver". It does not include GitBucket and Gitea: I'm not familiar with them and so I can't recommend (@ziotom78 , please do it, if you think they are worth). In general I would consider only the most popular solutions (GitHub, GitLab, BitBucket) but maybe it is only because I am too lazy to consider a solution that I haven't used so far.